fix: detach `FeedbackDatasetConfig` from `schemas.py`
