+++ src/argilla/client/feedback/schemas.py
@@ -460,58 +460,3 @@ class RankingQuestion(QuestionSchema):
         if isinstance(values.get("values"), list):
             values["settings"]["options"] = [{"value": value, "text": value} for value in values.get("values")]
         return values
-
-
-class FeedbackDatasetConfig(BaseModel):
-    """`FeedbackDatasetConfig`
-
-    Args:
-        fields (List[AllowedFieldTypes]): The fields of the feedback dataset.
-        questions (List[AllowedQuestionTypes]): The questions of the feedback dataset.
-        guidelines (Optional[str]): the guidelines of the feedback dataset. Defaults to None.
-
-    Examples:
-        >>> import argilla as rg
-        >>> config = rg.FeedbackDatasetConfig(
-        ...     fields=[
-        ...         rg.TextField(name="text", title="Human prompt"),
-        ...     ],
-        ...     questions =[
-        ...         rg.TextQuestion(
-        ...             name="question-1",
-        ...             description="This is the first question",
-        ...             required=True,
-        ...         ),
-        ...         rg.RatingQuestion(
-        ...             name="question-2",
-        ...             description="This is the second question",
-        ...             required=True,
-        ...             values=[1, 2, 3, 4, 5],
-        ...         ),
-        ...         rg.LabelQuestion(
-        ...             name="relevant",
-        ...             title="Is the response relevant for the given prompt?",
-        ...             labels=["Yes","No"],
-        ...             required=True,
-        ...             visible_labels=None
-        ...         ),
-        ...         rg.MultiLabelQuestion(
-        ...             name="content_class",
-        ...             title="Does the response include any of the following?",
-        ...             description="Select all that apply",
-        ...             labels={"cat-1": "Category 1" , "cat-2": "Category 2"},
-        ...             required=False,
-        ...             visible_labels=4
-        ...         ),
-        ...     ],
-        ...     guidelines="Add some guidelines for the annotation team here."
-        ... )
-
-    """
-
-    fields: List[AllowedFieldTypes]
-    questions: List[AllowedQuestionTypes]
-    guidelines: Optional[str] = None
-
-    class Config:
-        smart_union = True
